gcp-bookstore-data-project/review-publisher-app/requirements.txt

streamlit
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
pandas